Modified version of AVL Tree
I changed the method of searching for the depth/ height of a node. Instead of using a recursive method(which would be O(2^n)), I changed it so that the information is ready at any point. The method time taken to update the height/depth is O(log n). 
TLDR; It was slow, so I made it faster.